Basic Conversions Complete, Begin 'shinying up'
