+++ SwissCheese.vcxproj.filters
@@ -24,9 +24,6 @@
     <ClCompile Include="src\Calc.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\BaseConvert.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\SwissCheese.h">
@@ -35,13 +32,10 @@
     <ClInclude Include="src\Log.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\exprtk.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="src\Calc.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\BaseConvert.h">
+    <ClInclude Include="lib\exprtk.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
