--- src/BaseConvert.h
@@ -1,21 +1,0 @@
-#pragma once
-
-#include <iostream>
-#include <string>
-#include <bitset>
-#include <sstream>
-
-enum ConversionModes {
-	Binary,
-	Decimal,
-	Hexadecimal
-};
-
-struct ConversionParameters {
-	ConversionModes ModeA;
-	ConversionModes ModeB;
-	std::string InString;
-};
-
-void BinToHex();
-void BaseConversion(ConversionParameters InParams);
